requirements.txt
Django==3.1.4
django-extensions==3.1.0
django-model-utils==4.1.1
djangorestframework==3.12.2
